+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 All notable changes to this project will be documented in this file.
 
+
+## [0.4.4] - 2025-09-22
+
+* Add initializer generator
+* Add `ActiveJob::Uniqueness` support
+
 ## [0.4.3] - 2025-07-31
 
 * Remove job_options.discard_on` argument to `Readymade::BackgroundJob` to discard job on specific exceptions due to incorrect behaviour on different Rails versions